Docker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
transformers==4.37.0
torch==2.1.2
pillow==10.2.0
python-multipart==0.0.6
numpy<2
Pillow
